refactor: improve coin system and merge with landing score logic

+++ Assets/Scripts/LandingPad.cs
@@ -1,12 +1,9 @@
 using UnityEngine;
 
-public class LandingPad : MonoBehaviour
-{
+public class LandingPad : MonoBehaviour {
     [SerializeField] private int scoreMultiplier;
 
-
-    public int getScoreMultiplier()
-    {
+    public float getScoreMultiplier() {
         return scoreMultiplier;
     }
 }
